Implement some of C99 inttypes.h header (formatting macros for stdint.h). Use in appropriate places. Also replace use of 0x%lx with %p.

+++ uspace/app/mkfat/mkfat.c
@@ -43,6 +43,8 @@
 #include <mem.h>
 #include <devmap.h>
 #include <byteorder.h>
+#include <sys/types.h>
+#include <inttypes.h>
 #include <errno.h>
 #include "fat.h"
 
@@ -156,7 +158,8 @@ int main(int argc, char **argv)
 	if (rc != EOK) {
 		printf(NAME ": Warning, failed to obtain block device size.\n");
 	} else {
-		printf(NAME ": Block device has %llu blocks.\n", dev_nblocks);
+		printf(NAME ": Block device has %" PRIu64 " blocks.\n",
+		    (uint64_t) dev_nblocks);
 		cfg.total_sectors = dev_nblocks;
 	}
